feat(tree): add renderIcon props close #309
